requirements.txt
pandas
numpy
python-dateutil
mysql-connector-python
pandera
python-dotenv